add executive view template
